requirements.txt
Flask
yfinance
python-dotenv
gunicorn